Update materials.json
Added a recipe for White Paint using Zinc Sulfide as white pigment.  Found reference that mentioned rule of thumb in paint mixing is a ratio of 4:1 for medium to pigment.  Medium is a water based latex, creating 1L of paint base with 0.75L Water and 0.25L Latex, pigmented with 0.25L of Zinc Sulfide, to create the 1.25L of a 100 batch of white paint.